- System information session temporarily deleted (will be returned after rewriting the host side). - Partial implementation of the file transfer session. - Bringing the code style to Qt style. - Deleting unused code.

+++ source/base/logging.h
@@ -612,7 +612,7 @@ public:
     void operator&(std::ostream&) { }
 };
 
-using SystemErrorCode = DWORD;
+using SystemErrorCode = unsigned long;
 
 // Alias for ::GetLastError() on Windows and errno on POSIX. Avoids having to
 // pull in windows.h just for GetLastError() and DWORD.
